backend/pyproject.toml
[project]
name = "dvp-backend"
version = "0.1.0"
description = "Distributed Verification Platform backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.24.0",
    "SQLAlchemy[asyncio]>=2.0",
    "asyncpg>=0.27.0",
    "aiosqlite>=0.19.0",
    "alembic>=1.11.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "httpx>=0.27.0",
    "pytest>=8.4.0",
    "pytest-xdist>=3.2.0",
    "pytest-cov>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "aiosmtplib>=2.0.0",
    "click>=8.1.0",
    "rich>=13.0.0"
]

[project.scripts]
dvp = "cli:cli"
